+++ include/kernel/fs/dentry.h
@@ -108,9 +108,9 @@ typedef struct dir_ctx
      * @return `true` to continue iterating, `false` to stop.
      */
     bool (*emit)(dir_ctx_t* ctx, const char* name, ino_t number, itype_t type);
-    uint64_t pos;   ///< The current position in the directory, can be used to skip entries.
-    void* private;  ///< Private data that the filesystem can use to conveniently pass data.
-    uint64_t index; ///< An index that the filesystem can use for its own purposes.
+    size_t pos;    ///< The current position in the directory, can be used to skip entries.
+    void* private; ///< Private data that the filesystem can use to conveniently pass data.
+    size_t index;  ///< An index that the filesystem can use for its own purposes.
 } dir_ctx_t;
 
 /**
